Implemented menu CQRS

+++ BuberDinner.Contracts/Menus/MenuResponse.cs
@@ -4,7 +4,7 @@ public record MenuResponse(
     string Id,
     string Name,
     string Description,
-    float? AverageRating,
+    double? AverageRating,
     List<MenuSectionResponse> Sections,
     string HostId,
     List<string> DinnerIds,
